Implement tarball release process
The source distribution will be posted in the repo's Releases when tags
are pushed.

PR's labelled 'ignore-for-release' will not be released.

Discussions are automatically created when releases are made. See
https://docs.github.com/en/repositories/releasing-projects-on-github/automatically-generated-release-notes#configuring-automatically-generated-release-notes

Allows the build&test workflow to run on pushes to all branches and
all PR's, including generating the artifact containing a tarball.